Player Can Pick Up Bread
I created a script that allows the player to pick up a piece of bread when they touch a collider. After they pick up the bread, they can walk around with it as well. However, once the player picked up the bread, it was a clone. The clone was the original size of the object, and I had to change something in my code from false to true to change the size of the bread to match the player's size.