Stove Trigger
I added a box collider on the stove. I made sure that when the player collides with the box collider and clicks the space key, the word "Space" is printed in the console. Once I achieved this, the player was able to go through the stove and then fall through the floor. To solve this issue, I added a second box collider to the stove and made sure to disable the trigger box.