Place A Pan In The Scene
I placed a pan in the scene and duplicated a piece of bread. I placed the bread in the pan and made the bread invisible when the game starts. Once the games starts, the player can pick up a piece of bread, go near the stove, place the bread in the pan, and finally the newly, duplicated bread can appear. I also made sure to destroy the bread that is in the player's hands.