Added Eggs To The Game
I added eggs to the game, that the player can grab and bring to the stove. After coming to the stove, they can drop the egg onto the pan, by clicking the space key. This would make it seem as though the player is frying an egg. Of course, the egg looks like it was cracked as well.