Smoke Particle
I added a smoke particle to make it seem as though the food is actually cooking. I made sure to allow the player to be able to pick up the food from the stove, only once the timer is done. The smoke particle only works when the food is cooking as well.